Added stuff.

+++ ml4convection/process_radar_data.py
@@ -152,8 +152,9 @@ def _run(input_dir_name, first_date_string, last_date_string,
             output_file_name=this_netcdf_file_name
         )
 
-        radar_io.compress_file(this_netcdf_file_name)
-        os.remove(this_netcdf_file_name)
+        if os.path.isfile(this_netcdf_file_name):
+            radar_io.compress_file(this_netcdf_file_name)
+            os.remove(this_netcdf_file_name)
 
         if i != len(date_strings) - 1:
             print(SEPARATOR_STRING)
